requirements.txt
langchain==0.1.0
chromadb==0.4.22
pandas==2.1.4
openai==1.3.7
python-dotenv==1.0.0
SpeechRecognition==3.10.1
pyttsx3==2.90
PyAudio==0.2.13
sentence-transformers
